Move expression logic from RecursiveStatementVisitor to RecursiveAstVisitor
